Fixed integer field to cases, smallinteger was overflowing

+++ humming_brazil_covid19/report/models.py
@@ -69,8 +69,8 @@ class Case(models.Model):
     state = models.CharField(max_length=2, choices=STATES)
     region = models.CharField(max_length=2, choices=REGION)
 
-    cases = models.PositiveSmallIntegerField(default=0)
-    deaths = models.PositiveSmallIntegerField(default=0)
+    cases = models.PositiveIntegerField(default=0)
+    deaths = models.PositiveIntegerField(default=0)
 
     report = models.ForeignKey(Report, on_delete=models.CASCADE)
